refactor: split character management into modules
