+++ js/characters/render.js
@@ -1,0 +1,372 @@
+/* ===== Characters Management ===== */
+import { state, saveState } from '../state.js';
+import {
+  $,
+  slotsFromSTR,
+  speedFactorFromEmpty,
+  fmtSpeed,
+  slowdownLabel
+} from '../helpers.js';
+
+import { createSlot } from './slot.js';
+import { renderCharList } from './list.js';
+
+// Render characters in the center panel
+function renderChars() {
+  const wrap = $("#chars"); 
+  if (!wrap) return;
+  wrap.innerHTML = "";
+
+  // Initialize hiddenChars array if it doesn't exist
+  if (!Array.isArray(state.ui.hiddenChars)) {
+    state.ui.hiddenChars = [];
+  }
+
+  state.chars.forEach((c, ci) => {
+    // Skip rendering if character is hidden
+    if (state.ui.hiddenChars.includes(ci)) {
+      return;
+    }
+    // Initialize the equipped and backpack arrays if they don't exist
+    if (!c.equipped || !Array.isArray(c.equipped)) {
+      c.equipped = Array(9).fill(null);
+    }
+    
+    const backpackSlots = slotsFromSTR(c.str);
+    if (!c.backpack || !Array.isArray(c.backpack)) {
+      // If migrating from old format, transfer items to backpack
+      if (Array.isArray(c.slots)) {
+        c.backpack = c.slots.slice(0, backpackSlots);
+        delete c.slots; // Remove old slots array
+      } else {
+        c.backpack = Array(backpackSlots).fill(null);
+      }
+    }
+
+    // Ensure backpack array size matches STR rule
+    if (c.backpack.length !== backpackSlots) {
+      const old = c.backpack.slice(0, backpackSlots);
+      while (old.length < backpackSlots) old.push(null);
+      c.backpack = old; 
+      saveState();
+    }
+
+    // FIXED: Ensure equipped array is ALWAYS exactly 9 slots regardless of STR
+    // This fixes an issue where some characters might have less than 9 equipped slots
+    c.equipped = c.equipped || [];
+    if (c.equipped.length !== 9) {
+      const old = c.equipped.slice(0, Math.min(c.equipped.length, 9));
+      while (old.length < 9) old.push(null);
+      c.equipped = old;
+      saveState();
+    }
+    
+    // Check for sacks in equipped section
+    let hasSmallSack = false;
+    let hasLargeSack = false;
+    
+    // Look for sacks in equipped slots
+    for (let i = 0; i < c.equipped.length; i++) {
+      const item = c.equipped[i];
+      if (item && item.head) {
+        if (item.name === "Sack (small)") {
+          hasSmallSack = true;
+        } else if (item.name === "Sack (large)") {
+          hasLargeSack = true;
+        }
+      }
+    }
+    
+    // Initialize sack arrays if needed
+    if (hasSmallSack && (!c.smallSack || !Array.isArray(c.smallSack))) {
+      c.smallSack = Array(9).fill(null);
+    }
+    
+    if (hasLargeSack && (!c.largeSack || !Array.isArray(c.largeSack))) {
+      c.largeSack = Array(backpackSlots).fill(null);
+    }
+    
+    // Ensure sack arrays have the correct size
+    if (hasLargeSack && c.largeSack.length !== backpackSlots) {
+      const old = c.largeSack.slice(0, backpackSlots);
+      while (old.length < backpackSlots) old.push(null);
+      c.largeSack = old;
+      saveState();
+    }
+    
+    if (hasSmallSack && c.smallSack.length !== 9) {
+      const old = c.smallSack.slice(0, 9);
+      while (old.length < 9) old.push(null);
+      c.smallSack = old;
+      saveState();
+    }
+
+    // Count occupied slots in backpack and large sack (if equipped)
+    let totalSlots = backpackSlots;
+    let totalUsed = c.backpack.filter(x => !!x && (x.head || x.link !== undefined)).length;
+    
+    // Include large sack in encumbrance calculation if equipped
+    if (hasLargeSack && c.largeSack) {
+      totalSlots += backpackSlots; // Large sack has same capacity as backpack
+      totalUsed += c.largeSack.filter(x => !!x && (x.head || x.link !== undefined)).length;
+    }
+    
+    const empty = totalSlots - totalUsed;
+    
+    // Calculate backpack encumbrance factor
+    const backpackFactor = speedFactorFromEmpty(empty);
+    
+    // Calculate equipped-based speed penalties
+    let equippedSpeedFactor = 1.0; // Default is 100% speed (120')
+    
+    // Count filled equipped slots from top down
+    const filledEquippedSlots = c.equipped.filter(slot => slot !== null).length;
+    
+    // Apply speed penalties based on filled equipped slots
+    if (filledEquippedSlots >= 8) {
+      equippedSpeedFactor = 0.25; // 8-9 slots = speed 30
+    } else if (filledEquippedSlots >= 6) {
+      equippedSpeedFactor = 0.50; // 6-7 slots = speed 60
+    } else if (filledEquippedSlots >= 4) {
+      equippedSpeedFactor = 0.75; // 4-5 slots = speed 90
+    } // 1-3 slots = speed 120 (no penalty, equippedSpeedFactor remains 1.0)
+    
+    // Use the more restrictive of the two factors
+    const factor = Math.min(backpackFactor, equippedSpeedFactor);
+    const effFt = Math.max(0, Math.round(120 * factor)); // fixed base speed
+
+    const charEl = document.createElement("div");
+    charEl.className = "char";
+
+    let encClass = "enc-good";
+    if (empty < 6) encClass = "enc-warn";
+    if (empty < 4) encClass = "enc-warn";
+    if (empty < 2) encClass = "enc-bad";
+
+    // Display slots that include large sack if equipped
+    const displaySlots = hasLargeSack ? totalSlots : backpackSlots;
+    
+    charEl.innerHTML = `
+      <div class="char-header">
+        <div class="char-name">${c.name}</div>
+      <div class="char-meta">
+          <span class="pill">STR ${c.str}</span>
+          <span class="pill">Slots ${displaySlots}</span>
+          <span class="pill ${encClass}">${slowdownLabel(empty, totalSlots)}</span>
+          <span class="pill ${effFt >= 120 ? 'enc-good' : effFt >= 90 ? 'enc-warn' : effFt >= 60 ? 'enc-warn' : 'enc-bad'}">Speed ${fmtSpeed(effFt)}</span>
+        </div>
+      </div>
+    `;
+
+    const inv = document.createElement("div");
+    inv.className = "inv";
+
+    // Equipped section
+    const equippedSection = document.createElement("div");
+    equippedSection.className = "inventory-section";
+    
+    // Initialize equipped collapse state in state if it doesn't exist
+    if (!state.ui.equippedCollapsed) {
+      state.ui.equippedCollapsed = [];
+    }
+    
+    // Ensure array is large enough for this character
+    while (state.ui.equippedCollapsed.length <= ci) {
+      state.ui.equippedCollapsed.push(false);
+    }
+    
+    const equippedTitle = document.createElement("div");
+    equippedTitle.className = "section-title";
+    
+    // Create title with arrow collapse button on the left
+    equippedTitle.innerHTML = `
+      <button class="icon-btn collapse-btn" data-char="${ci}" data-section="equipped">
+        ${state.ui.equippedCollapsed[ci] ? "▶" : "▼"}
+      </button>
+      <span>Equipped</span>
+    `;
+    equippedSection.appendChild(equippedTitle);
+    
+    const equippedCol = document.createElement("div");
+    equippedCol.className = "slots";
+    
+    // Apply collapsed state if needed
+    if (state.ui.equippedCollapsed[ci]) {
+      equippedCol.style.display = "none";
+    }
+    
+    // Render equipped slots
+    c.equipped.forEach((slotObj, si) => {
+      const slot = createSlot(slotObj, ci, si, "equipped", c.equipped, backpackSlots, renderChars, renderCharList);
+      equippedCol.appendChild(slot);
+    });
+    
+    equippedSection.appendChild(equippedCol);
+    inv.appendChild(equippedSection);
+    
+    // Backpack section
+    const backpackSection = document.createElement("div");
+    backpackSection.className = "inventory-section";
+    
+    // Initialize backpack collapse state in state if it doesn't exist
+    if (!state.ui.backpackCollapsed) {
+      state.ui.backpackCollapsed = [];
+    }
+    
+    // Ensure array is large enough for this character
+    while (state.ui.backpackCollapsed.length <= ci) {
+      state.ui.backpackCollapsed.push(false);
+    }
+    
+    const backpackTitle = document.createElement("div");
+    backpackTitle.className = "section-title";
+    
+    // Create title with arrow collapse button on the left
+    backpackTitle.innerHTML = `
+      <button class="icon-btn collapse-btn" data-char="${ci}" data-section="backpack">
+        ${state.ui.backpackCollapsed[ci] ? "▶" : "▼"}
+      </button>
+      <span>Backpack</span>
+    `;
+    backpackSection.appendChild(backpackTitle);
+    
+    const backpackCol = document.createElement("div");
+    backpackCol.className = "slots";
+    
+    // Apply collapsed state if needed
+    if (state.ui.backpackCollapsed[ci]) {
+      backpackCol.style.display = "none";
+    }
+
+    // Create a helper function for creating slot elements
+
+    // Render backpack slots
+    c.backpack.forEach((slotObj, si) => {
+      const slot = createSlot(slotObj, ci, si, "backpack", c.backpack, backpackSlots, renderChars, renderCharList);
+      backpackCol.appendChild(slot);
+    });
+
+    backpackSection.appendChild(backpackCol);
+    inv.appendChild(backpackSection);
+    
+    // Add sack sections if equipped
+    // Check for large sack first (takes priority)
+    if (hasLargeSack) {
+      const largeSackSection = document.createElement("div");
+      largeSackSection.className = "inventory-section";
+      
+      // Initialize large sack collapse state in state if it doesn't exist
+      if (!state.ui.largeSackCollapsed) {
+        state.ui.largeSackCollapsed = [];
+      }
+      
+      // Ensure array is large enough for this character
+      while (state.ui.largeSackCollapsed.length <= ci) {
+        state.ui.largeSackCollapsed.push(false);
+      }
+      
+      const largeSackTitle = document.createElement("div");
+      largeSackTitle.className = "section-title";
+      
+      // Create title with arrow collapse button on the left
+      largeSackTitle.innerHTML = `
+        <button class="icon-btn collapse-btn" data-char="${ci}" data-section="largeSack">
+          ${state.ui.largeSackCollapsed[ci] ? "▶" : "▼"}
+        </button>
+        <span>Large Sack</span>
+      `;
+      largeSackSection.appendChild(largeSackTitle);
+      
+      const largeSackCol = document.createElement("div");
+      largeSackCol.className = "slots";
+      
+      // Apply collapsed state if needed
+      if (state.ui.largeSackCollapsed[ci]) {
+        largeSackCol.style.display = "none";
+      }
+      
+      // Render large sack slots
+      c.largeSack.forEach((slotObj, si) => {
+        const slot = createSlot(slotObj, ci, si, "largeSack", c.largeSack, backpackSlots, renderChars, renderCharList);
+        largeSackCol.appendChild(slot);
+      });
+      
+      largeSackSection.appendChild(largeSackCol);
+      inv.appendChild(largeSackSection);
+    }
+    // If no large sack but small sack is equipped
+    else if (hasSmallSack) {
+      const smallSackSection = document.createElement("div");
+      smallSackSection.className = "inventory-section";
+      
+      // Initialize small sack collapse state in state if it doesn't exist
+      if (!state.ui.smallSackCollapsed) {
+        state.ui.smallSackCollapsed = [];
+      }
+      
+      // Ensure array is large enough for this character
+      while (state.ui.smallSackCollapsed.length <= ci) {
+        state.ui.smallSackCollapsed.push(false);
+      }
+      
+      const smallSackTitle = document.createElement("div");
+      smallSackTitle.className = "section-title";
+      
+      // Create title with arrow collapse button on the left
+      smallSackTitle.innerHTML = `
+        <button class="icon-btn collapse-btn" data-char="${ci}" data-section="smallSack">
+          ${state.ui.smallSackCollapsed[ci] ? "▶" : "▼"}
+        </button>
+        <span>Small Sack</span>
+      `;
+      smallSackSection.appendChild(smallSackTitle);
+      
+      const smallSackCol = document.createElement("div");
+      smallSackCol.className = "slots";
+      
+      // Apply collapsed state if needed
+      if (state.ui.smallSackCollapsed[ci]) {
+        smallSackCol.style.display = "none";
+      }
+      
+      // Render small sack slots
+      c.smallSack.forEach((slotObj, si) => {
+        const slot = createSlot(slotObj, ci, si, "smallSack", c.smallSack, backpackSlots, renderChars, renderCharList);
+        smallSackCol.appendChild(slot);
+      });
+      
+      smallSackSection.appendChild(smallSackCol);
+      inv.appendChild(smallSackSection);
+    }
+    
+    // Add event listeners for collapse buttons
+    charEl.querySelectorAll('.collapse-btn').forEach(btn => {
+      btn.addEventListener('click', (e) => {
+        // Use currentTarget instead of target to ensure we get the button element
+        // even if a child element within the button was clicked
+        const charIndex = parseInt(e.currentTarget.dataset.char, 10);
+        const section = e.currentTarget.dataset.section;
+        
+        // Toggle the collapsed state
+        if (section === 'equipped') {
+          state.ui.equippedCollapsed[charIndex] = !state.ui.equippedCollapsed[charIndex];
+        } else if (section === 'backpack') {
+          state.ui.backpackCollapsed[charIndex] = !state.ui.backpackCollapsed[charIndex];
+        } else if (section === 'largeSack') {
+          state.ui.largeSackCollapsed[charIndex] = !state.ui.largeSackCollapsed[charIndex];
+        } else if (section === 'smallSack') {
+          state.ui.smallSackCollapsed[charIndex] = !state.ui.smallSackCollapsed[charIndex];
+        }
+        
+        saveState();
+        renderChars();
+      });
+    });
+    
+    charEl.appendChild(inv);
+  wrap.appendChild(charEl);
+  });
+}
+
+export { renderChars };
+
